Feat: implement Meet SDK hook
